move regfitworkspace to fitter.py
